Aggregate refactored
