+++ core/operation/src/main/java/uk/gov/gchq/gaffer/operation/impl/function/Aggregate.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.gov.gchq.gaffer.operation.function;
+package uk.gov.gchq.gaffer.operation.impl.function;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import org.apache.commons.lang3.exception.CloneFailedException;
